test: refactor existing tests and added tests for generate passphrase function
